# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Midanic E-Commerce & ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /auth/select-store:
    post:
      operationId: selectStore
      tags: [auth]
      summary: Switch to another store the user has access to (re-issues JWT)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectStoreRequest"
      responses:
        "200":
          description: Selected store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SelectStoreResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        preferredLang:
          type: string
          enum: [ar, en]
      required: [name, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
        stores:
          type: array
          items:
            $ref: "#/components/schemas/Store"
        currentStoreId:
          type: integer
          nullable: true
      required: [token, user]
    Store:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        slug:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, nameAr, nameEn, slug]
    SelectStoreRequest:
      type: object
      properties:
        storeId:
          type: integer
      required: [storeId]
    SelectStoreResponse:
      type: object
      properties:
        token:
          type: string
        currentStoreId:
          type: integer
        store:
          $ref: "#/components/schemas/Store"
      required: [token, currentStoreId, store]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, admin, employee]
        preferredLang:
          type: string
          enum: [ar, en]
        stores:
          type: array
          items:
            $ref: "#/components/schemas/Store"
        currentStoreId:
          type: integer
          nullable: true
      required: [id, name, email, role]

    # Products
...